% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames.R
\name{call_geonames}
\alias{call_geonames}
\title{Get latitude and longitude from GeoNames}
\usage{
call_geonames(
  location,
  country = "",
  continent = "",
  bypass_first_attempts = FALSE
)
}
\arguments{
\item{location}{A placename.}

\item{country}{A country name, useful for those towns appearing in multiple
countries, Default: ''}

\item{continent}{A continent name, useful for the same reason as the 
\code{country} parameter, Default: ''}

\item{bypass_first_attempts}{Logical. Whether to go straight to the fallback
Default: FALSE}
}
\value{
A named list \code{latitude, longitude)}, where each value is either
  a \code{double} or \code{NA}.
}
\description{
\code{call_geonames()} is a wrapper to call the Geonames database
  and retrieve latitude and longitude for a given location
}
\details{
\code{continent} parameter is ignored if a first attempt with it
  fails.
}
\examples{
\dontrun{
if(interactive()){
 call_geonames('Paris', 'France')
 call_geonames('Istanbul', 'Turkey', 'Asia')# Still works despite geonames placing it in Europe
 }
}
}
\seealso{
\code{\link[httr]{POST}}
 \code{\link[xml2]{read_xml}}
 \code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
}
