% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{download_event}
\alias{download_event}
\title{Download a webpage}
\usage{
download_event(url, be_nice = TRUE, delay = 1)
}
\arguments{
\item{url}{A character vector.}

\item{be_nice}{Whether to wait a bit after downloading, Default: TRUE}

\item{delay}{How long to wait, in seconds, Default: 1}
}
\value{
A \code{xml_document}
}
\description{
Download a webpage in a slightly more polite way. This function
  is meant to be lapplied.
}
\seealso{
\code{\link[xml2]{read_xml}}
}
