% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{find_more_codes}
\alias{find_more_codes}
\title{Convenience wrapper for \code{recurse_past_editions}}
\usage{
find_more_codes(event_codes)
}
\arguments{
\item{event_codes}{Event codes for which to guess past editions}
}
\value{
A character vector of "guessed" event codes
}
\description{
Apply \code{recurse_past_editions} for each \code{event_code}
  provided, make the results into a neat, NA-free character vector.
}
\seealso{
\code{\link[swingscrapeit]{recurse_past_editions}}
}
