% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_class.R, R/gathering.R
\docType{class}
\name{Event-class}
\alias{Event-class}
\alias{find_past_event_nodes_and_codes}
\title{A class to hold information about a dance event}
\usage{
find_past_event_nodes_and_codes(doc)
}
\arguments{
\item{doc}{An \code{xml_document} of the past-events page}
}
\value{
An \code{Event} class

A named list with two values: \code{event_nodes} and \code{event_codes},
  where \code{event_nodes} is a list of \code{xml_node} and \code{event_codes}
  is a character vector.
}
\description{
A (probably not optimal way) of holding all the information
  about a dance event

Extract the html nodes and event codes (necessary to
  rebuild the past events urls) from the past events page
}
\details{
There are some weirdness with events whose name starts with a number,
  so we get the \code{event_code} by manipulating the event's name.
}
\section{Slots}{

\describe{
\item{\code{event_code}}{Character. A unique identifier for the event}

\item{\code{name}}{Character.}

\item{\code{url}}{Character. The organisers' page for the event}

\item{\code{event_format}}{Factor}

\item{\code{location}}{Character. Town/City name}

\item{\code{country}}{Character.}

\item{\code{continent}}{Character.}

\item{\code{description}}{Character.}

\item{\code{styles}}{Vector of factors. The dances present at the event.}

\item{\code{teachers}}{List of character vectors. The sets of teachers at the event.}

\item{\code{teachers_confirmed}}{Logical.}

\item{\code{competitions_held}}{Logical.}

\item{\code{is_new_event}}{Logical.}

\item{\code{latitude}}{Numeric.}

\item{\code{longitude}}{Numeric.}

\item{\code{start_date}}{Date.}

\item{\code{end_date}}{Date.}
}}

\seealso{


\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
 \code{\link[stringr]{str_match}},\code{\link[stringr]{case}},\code{\link[stringr]{str_replace}}
}
