% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{get_seen_event_logical_vector}
\alias{get_seen_event_logical_vector}
\title{Check for already seen events}
\usage{
get_seen_event_logical_vector(event_codes, db_conn)
}
\arguments{
\item{event_codes}{A character vector}

\item{db_conn}{A connection to the database used. Should contain an "\code{event}" table}
}
\value{
A logical vector, where seen events are \code{TRUE}
}
\description{
Compare \code{event_codes} gathered with those already present in database.
}
\seealso{
\code{\link[DBI]{dbReadTable}},
 \code{\link[dplyr]{select}},
 \code{\link[tibble]{as_tibble}}
}
