% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_class.R
\name{make_event_row}
\alias{make_event_row}
\title{Construct a row for the database's 'event' table}
\usage{
make_event_row(x, ...)
}
\arguments{
\item{x}{An Event object}

\item{sqlite_types}{Whether to adapt types for an SQLite database}
}
\value{
A list with all relevant slots
}
\description{
Construct a row for the database's 'event' table
}
