% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_class.R
\docType{class}
\name{Event-class}
\alias{Event-class}
\alias{new_Event}
\title{A class to hold information about a dance event}
\value{
An \code{Event} class
}
\description{
A (likely non-optimal) way of holding all the information
  about a dance event
}
\section{Slots}{

\describe{
\item{\code{event_code}}{Character. A unique identifier for the event}

\item{\code{name}}{Character.}

\item{\code{url}}{Character. The organisers' page for the event}

\item{\code{event_format}}{Factor}

\item{\code{location}}{Character. Town/City name}

\item{\code{country}}{Character.}

\item{\code{continent}}{Character.}

\item{\code{description}}{Character.}

\item{\code{teacher_description}}{Character.}

\item{\code{styles}}{Vector of factors. The dances present at the event.}

\item{\code{teachers}}{List of character vectors. The sets of teachers at the event.}

\item{\code{teachers_confirmed}}{Logical.}

\item{\code{competitions_held}}{Logical.}

\item{\code{is_new_event}}{Logical.}

\item{\code{latitude}}{Numeric.}

\item{\code{longitude}}{Numeric.}

\item{\code{start_date}}{Date.}

\item{\code{end_date}}{Date.}
}}

