% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{recurse_past_editions}
\alias{recurse_past_editions}
\title{Guess past editions of an \code{event_code}}
\usage{
recurse_past_editions(s, idx = NULL, pad = FALSE)
}
\arguments{
\item{s}{A character vector. The \code{event_code} to find past editions for}

\item{idx}{An integer. The current index for which to create a "guessed"
\code{event_code}. Used only in the recursion, don't put anything
here, Default: NULL}

\item{pad}{Whether to pad \code{idx} with a \code{0} (in order to handle dates).
Used only in the recursion, don't put anything here, Default: FALSE}
}
\value{
A character vector of "guessed" events, or NA if nothing was found.
}
\description{
Swing Plan-It keeps past events on the website, but
  doesn't make them accessible through the homepage nor the past-events
  page. However they follow simple naming conventions: if you already have
  a url \code{https://www.swingplanit.com/event/hot-rhythm-holiday-6}, then
  you can try and access the urls for \code{hot-rhythm-holiday-5}, 
  \code{hot-rhythm-holiday-4}, etc, and there are good chances that those
  pages will still be up. This function recursively go through the likely
  candidates and record them.
}
\examples{
\dontrun{
if(interactive()){
 recurse_past_editions('hot-rhythm-holiday7')
 recurse_past_editions('my-balboa-weekend-2020')
 }
}
}
\seealso{
\code{\link[stringr]{str_match}},
 \code{\link[swingscrapeit]{find_more_codes}}
}
