% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{screen_scrape_homepage}
\alias{screen_scrape_homepage}
\title{Read the past-events page of Swing Plan-it using phantomjs}
\usage{
screen_scrape_homepage()
}
\value{
An HTML page of class \code{xml_document}
}
\description{
The \code{http://www.swingplanit.com/past-events} page has
  pretty weird HTML issues going on, so we read it using phantomjs
  in order to get all the information we need
}
\details{
The small JS script used to download the page is created on the fly
  in a temporary location, and so is the output download. No cleanup needed.  
  You NEED \code{phantomjs} installed on your machine, it is called using
  \code{system}
}
