% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gathering.R
\name{find_past_event_nodes_and_codes}
\alias{find_past_event_nodes_and_codes}
\title{Parse the Swing Plan-It past events page}
\usage{
find_past_event_nodes_and_codes(doc)
}
\arguments{
\item{doc}{An \code{xml_document} of the past-events page}
}
\value{
A named list with two values: \code{event_nodes} and \code{event_codes},
  where \code{event_nodes} is a list of \code{xml_node} and \code{event_codes}
  is a character vector.
}
\description{
Extract the html nodes and event codes (necessary to
  rebuild the past events urls) from the past events page
}
\details{
There are some weirdness with events whose name starts with a number,
  so we get the \code{event_code} by manipulating the event's name.
}
\seealso{
\code{\link[rvest]{html_nodes}},\code{\link[rvest]{html_text}}
 \code{\link[stringr]{str_match}},\code{\link[stringr]{case}},\code{\link[stringr]{str_replace}}
}
